tests/requirements.txt
docker==7.1.0
grpcio==1.73.1
grpcio-tools==1.73.1
pytest==8.4.1
requests==2.32.4
